fix compilation bug on MacOS
related issue: https://gitee.com/openharmony/ark_js_runtime/issues/I5BISH

+++ ecmascript/js_runtime_options.h
@@ -502,8 +502,8 @@ private:
         true,
         R"(if true trigger full gc, else trigger semi and old gc)"};
     PandArg<int> arkProperties_ {"ark-properties", GetDefaultProperties(), R"(set ark properties)"};
-    PandArg<size_t> gcThreadNum_ {"gcThreadNum", 7, R"(set gcThreadNum. Default: 7)"};
-    PandArg<size_t> longPauseTime_ {"longPauseTime", 40, R"(set longPauseTime. Default: 40ms)"};
+    PandArg<uint32_t> gcThreadNum_ {"gcThreadNum", 7, R"(set gcThreadNum. Default: 7)"};
+    PandArg<uint32_t> longPauseTime_ {"longPauseTime", 40, R"(set longPauseTime. Default: 40ms)"};
     PandArg<std::string> aotOutputFile_ {"aot-output-file",
         R"(aot_output_file.m)",
         R"(Path to AOT output file. Default: "aot_output_file.m")"};
